Emailing testing and cleaning (#385)
* Fix 4 4 cleaning files (#374)

* CircleCI Test Coverage

* Code Climate

eslint and brakeman are disabled.

* DQL cleaned, Tracking Repository cleaned

The DQL files of the PanelBundle were not used.

The TrackingRepositories have been cleaned as some methods were not
used.

* Separation StatisticsManager

The StatisticsManager needs to be refactored.

* Fix services

* Code Climate, cleaning tkg/cron, Buttons (#376)

* Fix 4 4 cleaning (#378)

* Code Climate, cleaning tkg/cron, Buttons

* StatisticsManager, camel case

* Cleaning from Code Climate

* Cleaning (#379)

* Cleaning

* Cleaning, composer json for file upload

* Testing TrackingBundle, cleaning (#381)

* Cleaning

* Cleaning, composer json for file upload

* Testing for TrackingBundle, additionnal cleaning

* Fix circleci php-zip

* CircleCI

* CircleCI, fix level3 affiliate as Statistics

* Fix 4 4 maintenance (#382)

* Cleaning

* Cleaning, composer json for file upload

* Testing for TrackingBundle, additionnal cleaning

* Fix circleci php-zip

* CircleCI

* CircleCI, fix level3 affiliate as Statistics

* Cleaning

* Fix the profile page (#383)

The edition page was not accessible.

Quick Testing has been added.

* Feature/4.4/deleteemail (#384)

* List and delete an email from the Participants

The new bundle ParticipantBundle has been created to look for an email
and list it.
The button on the right allows the suppression of an email.

It is useful for tests for clicks to leads. It is necesary to comply to
the RGPD.

This is relative to #322

* Testing with user Alice for Search Participant

* Fix StatisticsManager

* Create an emailing, testing and cleaning

The progress div has been taken off.
Basic testing has been created. (display the modal)

The second part is displayed. The cancelation does not create the
emailing.
This fixes the bug and closes #291

* Fix classes for Code Climate